+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# unreleased
+
+## Breaking changes
+
+- (layout): removed Page, Container, Content and Sidebar components
+
 # v2.0.0-dev.6
 
 - (Label): uppercase text can be disabled setting `uppercase` prop to false.
@@ -7,7 +13,6 @@
 
 - (Select) added select and selectfield components that uses expanderBase and expanderSelect
 - (Tooltip) removed arrow
-
 
 # v2.0.0-dev.4